+++ src/main/java/com/example/pet/repository/BoardRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-public interface BoardRepository extends JpaRepository<Board, Long> {
+public interface BoardRepository extends JpaRepository<Board, Integer> {
     @Query("SELECT b FROM Board b ORDER BY b.postId DESC")
     List<BoardListResponseDto> findAllDesc();
-    public Board findByPostId(Long postId);
+    public Board findByPostId(int postId);
 }
